Linux

+++ Test/src/FlyingPointLight.cpp
@@ -2,7 +2,7 @@
 
 #include "HodEngine/Game/src/Component/SceneComponent.h"
 
-#include <SDK/SDL/include/SDL.h>
+#include <SDL2/SDL.h>
 
 namespace HOD
 {
@@ -37,6 +37,6 @@ namespace HOD
 		Uint32 ticks = SDL_GetTicks();
 		float time = ticks / 1000.0f;
 
-		_sceneComponent->SetPosition(_sceneComponent->GetPosition() + glm::vec2(0.0f, 1.0f) * sin((float)time) * dt);
+		_sceneComponent->SetPosition(_sceneComponent->GetPosition() + glm::vec2(0.0f, 1.0f) * (float)sin((float)time) * dt);
 	}
 }
